entrega2/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.31.0
pydantic==2.6.0
pydantic-settings==2.1.0
redis==5.0.1
httpx==0.26.0
python-multipart==0.0.9
pytz==2024.1
websockets==12.0
firebase-admin==6.4.0
apscheduler==3.10.4
email-validator==2.1.1
passlib==1.7.4
bcrypt==3.2.2